alteracoes estruturais

+++ src/main/java/br/com/jsa/agendaapi/repository/UsuarioRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import br.com.jsa.agendaapi.model.Usuario;
 
 @Repository
-public interface UsuarioRepository extends CrudRepository<Usuario, Long> {
+public interface UsuarioRepository extends CrudRepository<Usuario, Integer> {
 
     public Optional<Usuario> findByUsuario(String usuario);
     public Optional<Usuario> findByEmail(String email);
